+++ scivision/io/reader.py
@@ -1,10 +1,10 @@
-import importlib
 import os
 from urllib.parse import urljoin
 
 import fsspec
 import yaml
 
+from .installer import install_package
 from .wrapper import PretrainedModel
 
 SCIVISION_YAML_CONFIG = ".scivision-config.yaml"
@@ -21,16 +21,6 @@ def _parse_config(path: os.PathLike) -> dict:
     assert "X" in config["prediction_fn"]["args"].keys()
 
     return config
-
-
-def _package_exists(config: dict) -> bool:
-    """Check to see whether a package exists."""
-    try:
-        importlib.import_module(config["import"])
-    except ModuleNotFoundError:
-        return False
-
-    return True
 
 
 def load_pretrained_model(
@@ -61,16 +51,7 @@ def load_pretrained_model(
 
     config = _parse_config(config_url)
 
-    # now check to see whether the package exists
-    if not _package_exists(config):
-        # NOTE(arl), here is where we could grab the repo and install it
-        install_str = config["url"]
-        if install_str.endswith(".git"):
-            install_str = install_str[:-4]
-
-        raise Exception(
-            "Package does not exist. Try installing it with: \n"
-            f"`!pip install -e git+{install_str}@main#egg={config['import']}`"
-        )
+    # try to install the package if necessary
+    install_package(config, allow_install=allow_install)
 
     return PretrainedModel(config)
